Move decorated SQLAlchemy UUID type into separate module.
